Extract MeshData from and place it in render_mesh.h
This simply pulls the implementation of the .h and .cc files. Things won't
work yet.
